requirements.txt
aiogram==3.15.0
APScheduler==3.10.4
dateparser==1.2.0
python-dotenv==1.0.1
aiosqlite==0.20.0
redis==5.2.1
pytest==8.3.4
pytest-asyncio==0.24.0
